Improved: Make a helper method for determining best transform details with and without experimental normalize support.
