+++ projects/dxt-lossless-transform-bc1/src/experimental/normalize_blocks/determine_optimal_transform.rs
