+++ projects/dxt-lossless-transform-cli/src/debug_bc1/benchmark_determine_best.rs
@@ -1,4 +1,4 @@
-use super::BenchmarkDetermineBestCmd;
+use super::{determine_best_transform_details_with_estimator, BenchmarkDetermineBestCmd};
 use crate::{
     debug::{
         benchmark_common::{
@@ -13,13 +13,7 @@ use crate::{
 };
 use core::time::Duration;
 use dxt_lossless_transform_api::DdsFormat;
-use dxt_lossless_transform_bc1::{
-    determine_optimal_transform::{determine_best_transform_details, Bc1EstimateOptions},
-    experimental::normalize_blocks::determine_best_transform::{
-        determine_best_transform_details_with_normalization, Bc1EstimateOptionsWithNormalization,
-    },
-    Bc1TransformDetails,
-};
+use dxt_lossless_transform_bc1::Bc1TransformDetails;
 use std::fs;
 
 /// Configuration for benchmark execution
@@ -202,42 +196,12 @@ unsafe fn run_determine_best_once(
         }
     };
 
-    // Determine the best transform details using the appropriate API
-    let result = if experimental_normalize {
-        // Use experimental API with normalization support
-        let transform_options = Bc1EstimateOptionsWithNormalization {
-            file_size_estimator: estimator,
-            test_normalize_options: true,
-        };
-
-        let experimental_details = determine_best_transform_details_with_normalization(
-            data_ptr,
-            len_bytes,
-            transform_options,
-        )
-        .map_err(|e| {
-            TransformError::Debug(format!(
-                "determine_best_transform_details_with_normalization failed: {e}"
-            ))
-        })?;
-
-        // Convert to standard struct for compatibility
-        Bc1TransformDetails {
-            decorrelation_mode: experimental_details.decorrelation_mode,
-            split_colour_endpoints: experimental_details.split_colour_endpoints,
-        }
-    } else {
-        // Use standard API without normalization
-        let transform_options = Bc1EstimateOptions {
-            file_size_estimator: estimator,
-        };
-
-        determine_best_transform_details(data_ptr, len_bytes, transform_options).map_err(|e| {
-            TransformError::Debug(format!("determine_best_transform_details failed: {e}"))
-        })?
-    };
-
-    Ok(result)
+    determine_best_transform_details_with_estimator(
+        data_ptr,
+        len_bytes,
+        estimator,
+        experimental_normalize,
+    )
 }
 
 /// Print file result with throughput measured in MiB/s for determine_best_transform_details benchmark
