+++ projects/dxt-lossless-transform-cli/src/debug_bc1/calc_compression_stats.rs
@@ -1,4 +1,4 @@
-use super::CompressionStatsCmd;
+use super::{determine_best_transform_details_with_estimator, CompressionStatsCmd};
 use crate::{
     debug::{
         calc_compression_stats_common,
@@ -16,13 +16,7 @@ use crate::{
 };
 use core::sync::atomic::{AtomicUsize, Ordering};
 use dxt_lossless_transform_api::DdsFormat;
-use dxt_lossless_transform_bc1::{
-    determine_optimal_transform::{determine_best_transform_details, Bc1EstimateOptions},
-    experimental::normalize_blocks::determine_best_transform::{
-        determine_best_transform_details_with_normalization, Bc1EstimateOptionsWithNormalization,
-    },
-    transform_bc1, Bc1TransformDetails,
-};
+use dxt_lossless_transform_bc1::{transform_bc1, Bc1TransformDetails};
 use dxt_lossless_transform_common::{allocate::allocate_align_64, color_565::YCoCgVariant};
 use rayon::iter::{IndexedParallelIterator, IntoParallelRefIterator, ParallelIterator};
 use std::{fs, sync::Mutex};
@@ -242,37 +236,12 @@ unsafe fn analyze_bc1_api_recommendation(
         }
     };
 
-    // Determine the best transform details using the appropriate API
-    let best_details = if experimental_normalize {
-        // Use experimental API with normalization support
-        let transform_options = Bc1EstimateOptionsWithNormalization {
-            file_size_estimator: estimator,
-            test_normalize_options: true,
-        };
-
-        let experimental_details = determine_best_transform_details_with_normalization(
-            data_ptr,
-            len_bytes,
-            transform_options,
-        )
-        .map_err(|e| {
-            TransformError::Debug(format!("Experimental API recommendation failed: {e}"))
-        })?;
-
-        // Convert to standard struct for compatibility
-        Bc1TransformDetails {
-            decorrelation_mode: experimental_details.decorrelation_mode,
-            split_colour_endpoints: experimental_details.split_colour_endpoints,
-        }
-    } else {
-        // Use standard API without normalization
-        let transform_options = Bc1EstimateOptions {
-            file_size_estimator: estimator,
-        };
-
-        determine_best_transform_details(data_ptr, len_bytes, transform_options)
-            .map_err(|e| TransformError::Debug(format!("API recommendation failed: {e}")))?
-    };
+    let best_details = determine_best_transform_details_with_estimator(
+        data_ptr,
+        len_bytes,
+        estimator,
+        experimental_normalize,
+    )?;
 
     // Transform the data using the recommended details and measure the size
     let mut transformed_data = allocate_align_64(len_bytes)?;
